Add stx-platform-helm RPM
Provide a new helm RPM that will be included with and installed by the
ISO.

This RPM will package an application tarball that is to be managed by
the system without user intervention but supports all the existing
mechanisms that the user installed applications use.

The intent of this application is to provide any charts required for
platform integration and install any of those charts through an armada
manifest.

The armada manifest will reference a new helm repository that will be
enabled with a future commit. This repository is intended for platform
specific charts installed by RPMs. The existing helm repository is
kept for optional runtime user installed applications.

Change-Id: Ia34d3a1efa1006b184cc35e17afb231064cf36f9
Story: 2005424
Task: 30452
Signed-off-by: Robert Church <robert.church@windriver.com>