Enable refreshing of helm repository information
This will automate execution of 'helm repo update' to ensure that the
helm repository information is up-to-date.

When the helm repos are updated via helm-upload and are re-indexed, the
local repository information also requires updating so that 'helm
search' commands report accurate information.

This will provide a utility function to perform this operation and the
utility function will be called after an application upload and when the
conductor is restarted.

This handles scenarios where an upload occurs on the active controller
followed by a swact. The newly actvated controller needs to make sure
that the local repository cache reflects any changes.

Change-Id: I67dba3f1a3a6e7c8169719ee622ddd533c69be31
Depends-On: Id2df178083961d46d069f3dc1590cb72a2cecd1b
Story: 2005424
Task: 30648
Signed-off-by: Robert Church <robert.church@windriver.com>