Change platform core pinning in AIO
The reason for this change is that the docker and kubernetes
processes are affined to the platform cores, and cpu
starvation can occur when those processes are loaded
with work, such as applying an armada manifest or downloading
docker images.

With this change, AIO will use the entire host cpu set rather than
just the platform cpu set.

Fixes-Bug: 1826592
Change-Id: Ic357e46804cac27b007fa58c52052970b2932780
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>