Set the number of nova schedulers to 1
When applying the stx-openstack manifest, the default
nova scheduler processes was equal to number of cores.

Now it is set to 1.

This reduces overall memory footprint and the time
to apply the stx-openstack manifest.  This reduces
out of memory issues on systems that have a very large
number of cores.

Change-Id: I7f9f41f13d5ae18da2c58d9968a7ee67cbe5a05e
Fixes-Bug: 1826308
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>