Bring up openstack services in stages during restore
During restore we need to bring up openstack services in stages.
First we bring up MariaDB to populate the database. Then we
bring up Glance and Cinder to restore images and volumes. Finally
Nova and other services will be brought up. This is achieved by
adding a "mode" option to sysinv application-apply API.

For the stx-openstack application the possible values are "restore_db",
"restore_storage" and "normal". The chartgroup list in the armada
manifest will be modified by the generated meta-overrides based on
the "mode".

Story: 2004761
Task: 30522

Change-Id: I2bdb8b3a9e1fe34ee6f5f6e376279a5dc14f7cc4
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>