Fix configuration failure when reinstalling worker node
When doing a system host-reinstall, the unlock of the worker would
fail with an error in the puppet logs.

This is because reinstall results in there not being any kubernetes
configuration, and therefore a kubernetes join is expected, however
the hieradata with that information was only being provided on an
initial creation of the node.

The fix was to always provide the token and command for the
kubernetes join command.

This fixes a host reinstall through sysinv.
(lock / reinstall / unlock)

An unsupported reinstall through a network boot (F12) will
typically fail since the token expires after 24 hours.
However that scenario can be remedied by a lock/unlock.

Change-Id: Idea9f1e4b8f98203260ec0b6af7ae29c34579b86
Fixes-Bug: 1823396
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>