Disable pre-allocation of 2M hugepages on VBox
Slow performance was observed in vbox with the default
2M hugepages allocation.

This update disables the default 2M hugepages allocation
in vbox. Users are expected to allocate desired huge pages
via CLI or Horizon.

In addition, the default platform reserved memory has been
increased to 2500M on the worker nodes running in vbox.

Fixes-Bug: 1825814

Change-Id: I2558bcde76a23f1293e07486dbbe5a4b6f91db2f
Signed-off-by: Tao Liu <tao.liu@windriver.com>