Enable CPU manager static policy for non-openstack worker nodes
For non-openstack based worker nodes the kubelet cpu-manager
static policy is enabled. The default 'none' policy is retained
for openstack based worker nodes and non-worker nodes.

This enables the complete CPU isolation of:
- Platform tasks (i.e., systemd tasks configured with CPUAffinity)
- kubernetes guaranteed pods (Gu)
- kubernetes besteffort (Be) and burstable (Bu) pods

Refer to following for CPU manager:
https://kubernetes.io/docs/tasks/administer-cluster/cpu-management-policies.

The platform reserved CPUs and memory are queried from sysinv database
for each worker node. Puppet configures the following cpu-manager
options specifically with 'static' policy:
--cpu-manager-policy=static
--system-reserved-cgroup=/system.slice
--system-reserved=cpu=<k8s_reserved_cpus>,memory=<k8s_reserved_mem>Mi

The Allocatable pool of resources is based on the total node capacity
and subtracting overheads including '--system-reserved' and
'--kube-reserved'. Specifying kube-reserved is not necessary.

Change-Id: I235c05b82d50bd9e3a8d36b0b076d8374dfb6055
Story: 2005192
Task: 29952
Signed-off-by: Jim Gauld <james.gauld@windriver.com>