fix static address assignment on interface network change
There are cases where changing the network assignments for an interface
results in an inconsistent set of static addresses for a host.  For
example, when setting multiple networks on a single interface it is
possible, depending on the order of the networks, that the api fails
to identify which interface is the host's mgmt interface.  This is
because of how the "networktype" attribute is used to identify
specific interfaces.  If the network has modified to add both the
"cluster-host", and "mgmt" networks, but the "cluster-host" network was
added first then that interface will not be identified as the "mgmt"
interface.

Since "networktype" is eventually going away, and since at the time
when that information is needed we already know the target interface,
there is no need to search (again) for the proper interface to use.

This is only problematic from the CLI if the networks are not added in
the expected order, but from the API any operation that adds both
networks at the same time may result in unexpected behaviour.

Change-Id: I86c8be4d0eb3b10b872b427bd05b3f5bfca4776a
Closes-Bug: 1827882
Signed-off-by: Allain Legacy <allain.legacy@windriver.com>