Add application apply prerequisites for platform managed apps
Add an application-apply dependency for the platform integration
application which launches the Ceph related charts. This dependency will
require that a quorum has been established and the crushmap has been
loaded prior to launching the application.

This will ensure that the charts have the Ceph connectivity required for
a successful chart release.

Change-Id: I56528200d16c68d129bc092e3dcc9af135cff16a
Story: 2005424
Task: 30977
Related-Bug: #1828896
Signed-off-by: Robert Church <robert.church@windriver.com>