Remove incorrect vswitch semantic check
We incorrectly required openstack compute label to unlock all worker
nodes if a vswitch_type was set. We allow mixed worker nodes on systems
with a vswitch type specified. ovs-dpdk is available for systems without
the openstack application, and for workers without the openstack compute
label.

This check has been removed.

Closes-Bug: 1828264
Change-Id: I16ca04cbcf5b4b8fcb484092ec9483fa38c9c031
Signed-off-by: David Sullivan <david.sullivan@windriver.com>