Enable PTP support for Link Aggregation
The Linux bonding driver aggregate multiple NICs into a single bonded
interface of two or more NIC slaves. This bonded interface has no
PTP hardware clock associated with it and cannot be used by the
LinuxPTP for clock synchronization. Need to derive all the LAG slaves
from the bonded interface and feed them to the LinuxPTP daemon instead
of the bonded interface. VLAN interface names should be cleared from
VLAN numbers as well in order for PTP to work.

Change-Id: I022e2d74e6459825129107881dd585bb2a13f953
Closes-bug: 1828017
Signed-off-by: Alex Kozyrev <alex.kozyrev@windriver.com>