Enable rbd-provisioner namespace overrides
This makes changes in two areas: the rbd-provisioner chart to allow
successful upgrades/rollbacks and the sysinv API for helm overrides so
that charts from all supported applications are displayed by the
helm-override-list command.

Additional updates to the rbd-provisioner chart allow the storage class
to be patched to overcome an armada error on upgrade and extending the
retry and active deadline to allow more tolerance if there are ceph
client related delays on install/upgrade/rollback.

Change-Id: Ib3e0169a52f86130ae32fc31c75cea01c4c67579
Story: 2005424
Task: 31066
Signed-off-by: Robert Church <robert.church@windriver.com>