Check patching api prior to uploading managed app
The patching operation check is happening too late when uploading a
platform managed application. If the patching API is not responding, the
managed application will be created in the database and end up stuck in
the uploading state as the upload will not proceed.

Move the check to the proper location to avoid prematurely creating the
application if it can't be loaded until patch dependencies are verified.

Also, prior to spawning the greenthread for application apply, save the
state to 'applying'. This will prevent multiple audit cycles from
starting greenthreads for the same action.

Change-Id: Iada51635a6965acb404be05cfd4d9ef167aebe9f
Closes-Bug: #1829791
Signed-off-by: Robert Church <robert.church@windriver.com>