Fix stuck config out of date alarm on worker nodes
This commit removes an argumment passed to the function that applies
the runtime puppet manifests. Problem with it is that we already pass
it in the config_dict and that it is very easy for developers to skip
this argument.

So, if we don't properly pass the list of hosts to
_config_apply_runtime_manifest it updates the puppet manifests for
all hosts of given personality when force=True is used, instead of
only doing it for the intended host.

This commit also makes sure that we do not apply reboot required
configs at runtime by making sure that config_uuid does not have
the reboot required flag set.

Change-Id: I378cfc798b18920b58c8e453a28ef040167416d1
Closes-Bug: 1830222
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>