Use true for force_raw_images when using ceph image backend
We need this patch for two reasons:

Nova of starlingx has not this patch[0]. We use remote storage(ceph)
as nova backend. If we set force_raw_image to False and use qcow2
format image to boot vms, the vms will fail to boot.

Nova of starlingx will have this patch[0]. If we still use False for
force_raw_images, the nova-compute service will refuse to start.

So that, we must set this force_raw_images to True at all.

[0] https://review.opendev.org/#/c/640271/

Change-Id: I76d48a4bf6846783edb4bb80724f777288bd9327
Story: 2003909
Task: 33547
Signed-off-by: zhu.boxiang <zhu.boxiang@99cloud.net>