Disable PCH sanity check on compute nodes (PTP slaves)
There can be configurations (i.e. LAG or VLAN) that use several NICs
with different PTP hardware clocks on management network.
boundary_clock_jbod allows such configuration and needs to be enabled.
Before we used this option on controller nodes only because of PTP
enabling on both OAM and management network interfaces. With LAG
support this option needs to be enabled on compute nodes as well.

Change-Id: I523844fce76bd2a90332d31e6104e4eaa36c6ca1
Closes-bug: 1828017
Signed-off-by: Alex Kozyrev <alex.kozyrev@windriver.com>