Remove extra system_mode in platform.conf
This commit removes the obsolete code that used to serve as a
workaround for a maintenance issue which has long been resolved.

The existence of this obsolete code has a side effect as
described in LP https://bugs.launchpad.net/starlingx/+bug/1829059
which is observed during bootstrap phase and in duplex systems
only.

The impact of this side effect has not been realized as the system
mode is persisted correctly in sysinv database and both simplex
and duplex systems have been brought up successfully with Ansible
bootstrap. However, this could cause some confusion to the users
who view the platform.conf file on controller-0 or potential
incorrect value to custom utility tools/processes that obtain this
parameter from platform.conf before the initial controller-0 unlock.

Test:
  Verify that /etc/platform/platform.conf on controller-0 in a
  standard system contains only one entry of system_mode
  (system_mode=duplex) after playbook run and and that the
  system can be brought up successfully.

Closes-Bug: 1829059
Change-Id: I7706e122ec28cf38d63c3bbcb00825125a32c76d
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>