Update rbd-provisioner replicas based on installed controllers
Currently the number of rbd-provisioner replicas is driven by the
stx-openstack application's 'openstack-control-plane' labels.

On systems where this label has not been applied to the controllers,
this will result in zero provisioners being installed.

Break the dependency on the stx-openstack app and set the number of
replicas based on the number of installed controllers as the
rbd-provisioner node selector will install in k8s masters (i.e.
controllers).

Also update the provisioner's storage-init pod to align with the same
node selection criteria as the rbd-provisioner pod.

Change-Id: Ida180fd12a4923c8cdd5bccf25a1a1e2af4f8a90
Closes-Bug: #1830290
Signed-off-by: Robert Church <robert.church@windriver.com>