Fix: SR-IOV device plugin config file mount
This commit fixes a bug that may occur on the second node of an
AIO-DX system if the sriovdp=enabled label has been applied after
initial install but before first unlock.

In this case, the device plugin may start launching before the
worker manifest has populated the config file.  Since the pod
attempts to mount the config file and it is not present, it gets
created as a directory.

The fix simply sets the hostPath type for the config file to 'File'

Closes-Bug: #1831124
Change-Id: Ie55a0cb80dcee81a81059118ebef6c8c01cb6996
Signed-off-by: Steven Webster <steven.webster@windriver.com>