Remove the is_kubernetes checks from sysinv
The is_kubernetes checks were added when containerization
was first being implemented to allow a non containerized
environment to continue to work.
The non kubernetes environment is no longer operational
and the option to disable this is no longer provided, so all
the code that checks for this system capability is no longer
needed, since it will always be True.

Change-Id: Iae91d6cb6f9a4fe19fad992df318b4cd609c6c97
Story: 2004764
Task: 33417
Depends-On: https://review.opendev.org/#/c/663779
Depends-On: https://review.opendev.org/#/c/663780
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>