Fix controllerfs-modify related bugs
A bug in the get_alarms_degrade was causing the API to return a list of
degrade affecting alarms when there was none. This could be observed by
raising NTP alarms by configuring the NTP servers with invalid
addresses. This bug would cause the controllerfs-modify to fail
increasing the size a filesystem.

Also fix a bug with the return value of the controllerfs-modify API. On
success "None" would be returned on the CLI.

Closes-Bug: 1828097
Change-Id: Id09652a8f88915ea4e5ec90c96f195fe0350a550
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>