Support Helm user overrides for generic applications
This commit updates to support the user overrides for generic
apps. The list of charts for generic apps is retrieved from
manifest and stored to database when uploading the app. The
application-upload and apply are changed to generate overrides
files for generic apps as well.

This also fixes the helm-override-list namespace display for
garbd, memcached and ingress charts in stx-openstack app

Tested(generic apps, stx-openstack app):
-system application-upload
-system helm-override-list
-system helm-override-update(--set --values)
-system helm-override-show
-system application-apply
-system application-remove
-system helm-override-delete

Change-Id: I45c561400fb82a69d51118159fad7bf03881f263
Story: 2003908
Task: 28631
Signed-off-by: Angie Wang <angie.wang@windriver.com>